+++ TreePhysics/Physics/UpdateRigidBodiesKernel.swift
@@ -52,7 +52,7 @@ final class UpdateRigidBodiesKernel: MetalKernel {
         let buffer = device.makeBuffer(length: count * MemoryLayout<JointStruct>.stride, options: [.storageModeShared])!
         let jointStructs = UnsafeMutableRawPointer(buffer.contents()).bindMemory(to: JointStruct.self, capacity: count)
         for i in 0..<count {
-            jointStructs[i] = JointStruct(θ: float3x3(0), k: 200)
+            jointStructs[i] = JointStruct(θ: half3x3(0), k: 200)
         }
         return buffer
     }
